fixed the home page controller and removed a console log
